backend/requirements/base.txt
Django==5.2.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.*
django-environ==0.11.*
django-cors-headers==4.6.*
mysqlclient==2.2.*
Pillow==11.*
gunicorn==23.*
